WIP: Bump mdsal to 5.0.0
This bumps mdsal to 5.0.0 as part of Magnesium MRI.

The change in uint{8,16,32,64} mapping necessitates quite a few
changes to protocol coding on both protocol-api and plugin-api
side, which are done in minimalistic fashion.

+++ openflowplugin-impl/src/main/java/org/opendaylight/openflowplugin/impl/datastore/multipart/PortDescMultipartWriter.java
@@ -24,6 +24,7 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.protocol.rev130731
 import org.opendaylight.yang.gen.v1.urn.opendaylight.port.statistics.rev131214.FlowCapableNodeConnectorStatisticsData;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.port.statistics.rev131214.FlowCapableNodeConnectorStatisticsDataBuilder;
 import org.opendaylight.yangtools.yang.binding.InstanceIdentifier;
+import org.opendaylight.yangtools.yang.common.Uint32;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -48,7 +49,7 @@ public class PortDescMultipartWriter extends AbstractMultipartWriter<MultipartRe
     public void storeStatistics(final MultipartReplyPortDesc statistics, final boolean withParents) {
         statistics.getPorts()
             .forEach(stat -> {
-                long portNumber = OpenflowPortsUtil.getProtocolPortNumber(
+                Uint32 portNumber = OpenflowPortsUtil.getProtocolPortNumber(
                         OpenflowVersion.get(features.getVersion()),
                         stat.getPortNumber());
                 final NodeConnectorId id = InventoryDataServiceUtil
