+++ openflowplugin-impl/src/test/java/org/opendaylight/openflowplugin/impl/device/initialization/OF13DeviceInitializerTest.java
@@ -48,6 +48,7 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.protocol.rev130731
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.protocol.rev130731.multipart.reply.multipart.reply.body.MultipartReplyDescCaseBuilder;
 import org.opendaylight.yangtools.yang.binding.KeyedInstanceIdentifier;
 import org.opendaylight.yangtools.yang.common.RpcResultBuilder;
+import org.opendaylight.yangtools.yang.common.Uint32;
 
 @RunWith(MockitoJUnitRunner.class)
 public class OF13DeviceInitializerTest {
@@ -97,7 +98,7 @@ public class OF13DeviceInitializerTest {
         when(deviceContext.getDeviceInfo()).thenReturn(deviceInfo);
         when(deviceContext.getMessageSpy()).thenReturn(messageSpy);
         when(deviceContext.oook()).thenReturn(translatorLibrary);
-        when(requestContext.getXid()).thenReturn(new Xid(42L));
+        when(requestContext.getXid()).thenReturn(new Xid(Uint32.valueOf(42L)));
         when(deviceContext.createRequestContext()).thenReturn(requestContext);
         when(deviceContext.getPrimaryConnectionContext()).thenReturn(connectionContext);
     }
